Move highlight state into <Control> component
